requirements.txt
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Django==3.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
future==0.18.2
idna==2.8
importlib-metadata==1.3.0
lxml==4.4.2
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pycparser==2.19
pyparsing==2.4.5
PySocks==1.7.1
pytest==5.3.2
pytest-django==3.8.0
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tornado==6.0.3
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
